dreamBot/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
backports.zoneinfo==0.2.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
discord.py==2.3.2
exceptiongroup==1.1.3
fastapi==0.101.1
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httpx==0.24.1
hupper==1.12
idna==3.4
logger==1.4
multidict==6.0.4
psycopg2-binary==2.9.7
pydantic==2.2.1
pydantic_core==2.6.1
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.23.2
yarl==1.9.2